refactor: merge mcp server into bridge component
