+++ Makefile
@@ -29,23 +29,14 @@ VENV       := bridge/venv
 PIP        := $(VENV)/bin/pip
 PYTHON     := $(abspath $(VENV)/bin/python)
 
-MCP_DIR    := $(abspath mcp_server)
-MCP_BIN    := mcp_server/dist/barbatos-mcp
-MCP_VENV   := mcp_server/venv
-MCP_PIP    := $(MCP_VENV)/bin/pip
-MCP_PYTHON := $(abspath $(MCP_VENV)/bin/python)
-
-.PHONY: install_dependencies compile_all compile_bridge_agent compile_bridge compile_mcp compile_binary prepare_release release run run_arm64_qemu
+.PHONY: install_dependencies compile_all compile_bridge_agent compile_bridge compile_binary prepare_release release run run_arm64_qemu
 
 install_dependencies:
-	rm -rf $(VENV) $(MCP_VENV)
+	rm -rf $(VENV)
 	python3 -m venv $(VENV)
 	$(PIP) install --upgrade pip
 	$(PIP) install -r bridge/requirements.txt
 	cd bridge && npm ci
-	python3 -m venv $(MCP_VENV)
-	$(MCP_PIP) install --upgrade pip
-	$(MCP_PIP) install -r mcp_server/requirements.txt
 
 compile_bridge_agent:
 	cd bridge && npm ci
@@ -55,13 +46,10 @@ compile_bridge_agent:
 compile_bridge: compile_bridge_agent
 	cd bridge && $(PYTHON) -m PyInstaller bridge.spec
 
-compile_mcp:
-	cd mcp_server && $(MCP_PYTHON) -m PyInstaller mcp.spec
-
 compile_binary:
 	./gradlew $(GRADLE_TARGET)
 
-compile_all: compile_binary compile_bridge_agent compile_bridge compile_mcp
+compile_all: compile_binary compile_bridge_agent compile_bridge
 
 release: compile_all prepare_release
 
@@ -75,6 +63,5 @@ prepare_release:
 	mkdir -p dist
 	cp $(TUI_BIN) dist/barbatos
 	cp $(BRIDGE_BIN) dist/barbatos-bridge
-	cp $(MCP_BIN) dist/barbatos-mcp
-	chmod +x dist/barbatos dist/barbatos-bridge dist/barbatos-mcp
+	chmod +x dist/barbatos dist/barbatos-bridge
 	@echo "Binaries ready in dist/"
