+++ bridge/bridge.spec
@@ -14,7 +14,7 @@ a = Analysis(
         ('agent.objc.bundle.js', '.'),
         ('../version.txt', '.')
     ] + frida_datas,
-    hiddenimports=['jdwp_frida', 'ios_repacker'] + frida_hidden,
+    hiddenimports=['jdwp_frida', 'ios_repacker', 'mcp'] + frida_hidden,
     hookspath=[],
     hooksconfig={},
     runtime_hooks=[],
