+++ bridge/requirements.txt
@@ -1,3 +1,4 @@
 frida==17.9.1
 frida-tools>=12.0.0
 pyinstaller>=6.0.0
+mcp
