--- mcp_server/mcp.spec
@@ -1,42 +1,0 @@
-# -*- mode: python ; coding: utf-8 -*-
-from PyInstaller.utils.hooks import collect_data_files, collect_submodules
-
-httpx_hidden = collect_submodules('httpx')
-httpx_datas = collect_data_files('httpx')
-
-a = Analysis(
-    ['server.py'],
-    pathex=[],
-    binaries=[],
-    datas=[('../version.txt', '.')] + httpx_datas,
-    hiddenimports=httpx_hidden,
-    hookspath=[],
-    hooksconfig={},
-    runtime_hooks=[],
-    excludes=[],
-    noarchive=False,
-    optimize=0,
-)
-
-pyz = PYZ(a.pure)
-
-exe = EXE(
-    pyz,
-    a.scripts,
-    a.binaries,
-    a.datas,
-    [],
-    name='barbatos-mcp',
-    debug=False,
-    bootloader_ignore_signals=False,
-    strip=False,
-    upx=False,
-    upx_exclude=[],
-    runtime_tmpdir=None,
-    console=True,
-    disable_windowed_traceback=False,
-    argv_emulation=False,
-    target_arch=None,
-    codesign_identity=None,
-    entitlements_file=None,
-)
