--- packaging/debian/barbatos-mcp/DEBIAN/control
@@ -1,8 +1,0 @@
-Package: barbatos-mcp
-Version: 1.0.0
-Section: utils
-Priority: optional
-Architecture: all
-Depends: python3, python3-mcp
-Maintainer: Victor Gazolli <victorlpgazolli@gmail.com>
-Description: MCP server for Barbatos debugger
